requirements.txt
# 请在终端（确保激活了 face_app 环境）执行下面这条长命令：
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu121

# --- 核心图像处理与算法 ---
# 锁定 numpy 1.26.4 是为了兼容 OpenCV 和 MediaPipe，且不与 PyTorch 2.4 冲突
numpy==1.26.4

# 面部识别与图像处理
opencv-python==4.9.0.80
mediapipe==0.10.9
pillow>=10.2.0

# --- 数据处理 ---
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.2

# --- 工具与可视化 ---
tqdm>=4.66.0
openpyxl>=3.1.2
tensorboard>=2.15.0
matplotlib>=3.8.0

# --- AIGC 生成相关 (为第三阶段做准备) ---
# diffusers 是 huggingface 的生图库，配合 A6000 飞快
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.26.0